+++ R/map_layer_text.R
@@ -84,6 +84,7 @@ add_text <- function(
 	text,
 	lon = NULL,
 	lat = NULL,
+	elevation = NULL,
 	polyline = NULL,
 	fill_colour = NULL,
 	fill_opacity = NULL,
@@ -120,6 +121,7 @@ add_text <- function(
 	l <- list()
 	l[["lon"]] <- force( lon )
 	l[["lat"]] <- force( lat )
+	l[["elevation"]] <- force( elevation )
 	l[["fill_colour"]] <- force( fill_colour )
 	l[["fill_opacity"]] <- resolve_opacity( fill_opacity )
 	l[["size"]] <- force( size )
@@ -168,7 +170,7 @@ add_text <- function(
 		geometry_column <- c( "geometry" )
 		shape <- rcpp_text_geojson( data, l, geometry_column, digits )
 	} else if ( tp == "df" ) {
-		geometry_column <- list( geometry = c("lon", "lat") )
+		geometry_column <- list( geometry = c("lon", "lat", "elevation") )
 		shape <- rcpp_text_geojson_df( data, l, geometry_column, digits )
 	} else if ( tp == "sfencoded" ) {
 		geometry_column <- "polyline"